[HACK] .github: Export the GLOO_IMAGE_REPO
This commit exports the GLOO_IMAGE_REPO in the GHA workflow
so the Makefile correctly set this. Necessarily to easily
test in my fork before introducing this into the repository
and/or before the cleanup stream of work has been pushed to
main and the gloo container image has been renamed.

Signed-off-by: timflannagan <timflannagan@gmail.com>