install/helm: Move service account config under controller
Move the service account configuration from the root level to be nested
under the controller section in values.yaml, as it's specifically tied
to the controller's operation.

This commit includes the following changes:

- Move serviceAccount config block under controller in values.yaml
- Update serviceaccount.yaml template to reference new config path
- Add descriptive comments to the serviceAccount config section
- Maintain consistent templating for ClusterRole references
- Fixes a bug where overriding the service account name results in
  broken installation

This change better organizes the Helm chart configuration by grouping
related settings together, making it clearer that the service account
is specifically for the controller component.

Closes https://github.com/kgateway-dev/kgateway/issues/10950

Signed-off-by: timflannagan <timflannagan@gmail.com>