WIP: Skip policy validation for non-spec updates
Overall, seems to be working where metadata only updates to a policy CR
that's already reporting an invalid state results in IR validation to
be skipped. Similarly, when an HTTPRoute's generation is bumped, I see
evidence that triggers another update to the policy collection and we
skip IR re-validation again.

However, the biggest issue I'm seeing now is that Pod restarts aren't
deterministic and result in status flipping between valid and invalid
despite the policy being flagged by the validation logic. Unclear
whether this a race condition or something else with status reporting.

Signed-off-by: timflannagan <timflannagan@gmail.com>