Attempt to fix the python unit test target, seems half baked, etc
The updated Go unit test job is online post-refactor, but the updated
python unit test job is broken as we shell out to the python/Makefile
and run a target that doesn't seem maintained.

Updated the target to call the AI extension unit tests instead and
handle the process of installing any dependencies. Note, not sure
if the latter is correct, just want to validate that CI is back
online with this change, but I cannot run that target successfully
using python3 in my local env.

Signed-off-by: timflannagan <timflannagan@gmail.com>