ir: Fix Service updates not propagating through translation
This change fixes an issue where patching Kubernetes Services (e.g.
changing appProtocol or ports) did not dynamically propagate to
the xDS config, and required restarting the kgw container.

The root cause was HttpRouteIR.rulesEqual() not comparing
BackendObject fields when checking backend equality. It was only
comparing ClusterName and Weight, missing changes to the underlying
Service configuration like appProtocol.

Implements an Equals() method for the HttpBackendOrDelegate type to
fix this edge case and cleanup the overall implementation to be
consistent with the rest of the codebase and avoid harcoding the
conditional checks related to Equals-based collection checking.

Fixes #12576.

Signed-off-by: timflannagan <timflannagan@gmail.com>