.github/actions: Optimize job for freeing disk space
Optimize the current GHA job that's responsible for freeing up
disk space for several runners. Use a custom, rust-based GHA
action that does this work instead of an inline bash script.
This effectively ports over a commit that Chandler made to
a repository outside of this org, but has significant performance
improvements.

Signed-off-by: timflannagan <timflannagan@gmail.com>
Co-authored-by: David L. Chandler <david.chandler@solo.io>
Signed-off-by: timflannagan <timflannagan@gmail.com>