ci: Automate release notes
Follow-up to the #13133 PR which introduced a bash script responsible
for generating custom release notes. This builds on top of that work
and hooks that script up to the rest of the release pipeline. Previously,
maintainers had to manually generate release notes and then edit the
release body after the workflow ran. Now, this branch introduces a
standalone job for changelog generation and leverages the goreleaser
--release-notes CLI flag to pipe the generated markdown file to the
goreleaser job in the release workflow.

Introduced this as a standalone job to make better use of parallelism
instead of adding another step in the goreleaser job. Additionally, this
workflow can still be tested in forks, however the data that gets scraped
from the script is using the kgateway-dev/kgateway (i.e. upstream) remote.

Tested in my fork: https://github.com/timflannagan/kgateway/releases/tag/v2.2.0-beta.4.

Fixes #13129.

Signed-off-by: timflannagan <timflannagan@gmail.com>